Fix file paths after moving files
