+++ client/js/verification-portal/index.js
@@ -5,7 +5,7 @@ import HeaderSuccess from './components/header-success'
 import Spacer from '../components/spacer'
 import LoginButton from '../authentication/facebook-login-button'
 import Tabs from './components/tabs'
-import LeftSideNavContainer from './components/left-side-nav-container'
+import LeftSideNavContainer from './left-side-nav-container'
 import CommunityHealthCenterForm from './community-health-center'
 import PregnancyResourceCenterForm from './pregnancy-resource-center'
 import { formTypes } from './constants'
